% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakString.R
\name{breakString}
\alias{breakString}
\title{Break a string}
\usage{
breakString(x, nChar = 20)
}
\arguments{
\item{x}{A string.}

\item{nChar}{The number of characters after which the new line is inserted. Default to 20.}
}
\value{
A string with inserted \code{\\n}.
}
\description{
The function breaks a string after around the specified number of characters.
}
\examples{
someText <- "This is the function that breaks a string."
breakString(x = someText, nChar = 20)
}
\author{
Marjan Cugmas
}
