% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFactorLabels.R
\name{makeFactorLabels}
\alias{makeFactorLabels}
\title{Make factor labels}
\usage{
makeFactorLabels(x, reduce = TRUE, ...)
}
\arguments{
\item{x}{Data for the selected variable, see Details.}

\item{reduce}{Wheter to reduce categories with zero frequency, default is \code{TRUE}.}

\item{...}{Arguments passed to function \code{factor}.}
}
\description{
The function transforms a numeric varibale into categorical one, based on the attribute data from a given SPSS file.
}
\details{
Data have to be imported by using the \code{foreign::read.spss} function.
The use of the function make sence when the parameter \code{use.value.lables} in the function \code{read.spss} is set to \code{FALSE}.
}
\author{
Aleš Žiberna
}
