% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_nominal_ordinal.R
\name{print.orc_result}
\alias{print.orc_result}
\title{Print method for ORC results}
\usage{
\method{print}{orc_result}(x, digits = NULL, ...)
}
\arguments{
\item{x}{Object returned by [orc_table()].}

\item{digits}{Number of digits used in rounded summary output. NULL (default) uses R defaults.}

\item{...}{Ignored.}
}
\description{
Print method for ORC results
}
