% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printCorTestDf.R
\name{printCorTestDf}
\alias{printCorTestDf}
\title{Print the results of the \code{CorTestDf} function}
\usage{
printCorTestDf(l, digits = c(3, 3), format = NULL)
}
\arguments{
\item{l}{Output of \code{corTestDf} function.}

\item{digits}{Vector of length two for the number of digits (the first element of a vector corresponds to the number of digits for correlation coeficients and the second element of a vector corresponds to the number of digits for \eqn{p}-values).}

\item{format}{a vector of length two for the formatting of the oputput values.}
}
\description{
The function prints the results of the \code{CorTestDf} function.
}
\examples{
corCars <- corTestDf(mtcars[, 3:5])
printCorTestDf(corCars, digits = c(2, 2))
}
\seealso{
\code{CorTestDf}
}
\author{
Aleš Žiberna
}
