% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameVar.R
\name{renameVar}
\alias{renameVar}
\title{Rename variables}
\usage{
renameVar(data, renames)
}
\arguments{
\item{data}{A dataframe.}

\item{renames}{A list with oldnames and newnames (e.g, \code{list("oldname1" = "newname1", "oldname2" = "newname2")}).}
}
\description{
The function for renaming one or several variables in a dataframe.
}
\examples{
renameVar(mtcars, list("cyl" = "Cylinders", "wt" = "Weight", "am" = "Transmission"))
}
\author{
Marjan Cugmas
}
