% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theta.R
\name{Theta}
\alias{Theta}
\title{Theta coefficient - measure of measurement internal consistency based on principal component analysis}
\usage{
Theta(C)
}
\arguments{
\item{C}{Covariance or correlation matrix.}
}
\value{
The value of the theta coefficient.
}
\description{
The function theta coefficient, which is a measure of measurement internal consistency based on principal component analysis, or more precisely first eigenvalue.
}
\examples{
Theta(C=cor(mtcars[,1:6]))
}
\author{
Ales Ziberna
}
