% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTestDf.R
\name{corTestDf}
\alias{corTestDf}
\title{Compute correlations and test their statistical significance}
\usage{
corTestDf(X, method = "p", use = "everything", ...)
}
\arguments{
\item{X}{Data matrix with selected variables.}

\item{method}{A type of correlation coefitient to be calculated, see function \code{cor}.}

\item{use}{In the case of missing values, which method should be used, see function \code{cor}.}

\item{...}{Arguments passed to other functions, see \code{cor.test}.}
}
\description{
The function computes the whole correlation matrix and corresponding sample sizes and \eqn{p}-values.
}
\examples{
corTestDf(mtcars[, 3:5])
}
\seealso{
\code{cor}
}
\author{
Aleš Žiberna
}
