% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTab.R
\name{freqTab}
\alias{freqTab}
\title{Create a frequency table}
\usage{
freqTab(x, dec = NULL, ...)
}
\arguments{
\item{x}{Vector with the values of a categorical variable.}

\item{dec}{Number of decimal places for percentages.}

\item{\dots}{Arguments passed to function \code{table}.}
}
\description{
The function creates a frequency table with percentages for the selected categorical variable.
}
\examples{
freqTab(mtcars[,2], dec = 1)
}
\author{
Aleš Žiberna
}
