% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTab.R
\name{freqTab}
\alias{freqTab}
\title{Create a frequency table}
\usage{
freqTab(x, dec = NULL, ...)
}
\arguments{
\item{x}{Vector with the values of a categorical variable.}

\item{dec}{Number of decimal places for percentages.}

\item{...}{Arguments passed to other functions, see \code{table}.}
}
\description{
The function creates a frequency table with percentages for the selected categorical variable.
}
\examples{
frekTab(GSSvocab[,4], dec = 1)
}
\author{
Aleš Žiberna
}
