% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printP.R
\name{printP}
\alias{printP}
\title{Print p-value}
\usage{
printP(p)
}
\arguments{
\item{p}{Value to be printed.}
}
\description{
The function round and prints \code{p}-value.
}
\examples{
printP(p = 0.523)
printP(p = 0.022)
printP(p = 0.099)
}
\author{
Marjan Cugmas
}
