% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wardKF.R
\name{wardKF}
\alias{wardKF}
\title{Calculate the value of the Ward criterion function}
\usage{
wardKF(X, clu)
}
\arguments{
\item{X}{Data frame with values of numberical variables (usually the ones that were/are used for clustering).}

\item{clu}{Partition.}
}
\value{
The value of the Ward criterion function.
}
\description{
The function calculate the value of the Ward criterion function, based on a set of numberical variables and one categorical variable (partition).
}
\author{
Aleš Žiberna
}
